requirements.txt
Pillow==7.2.0
potracer==0.0.1
numpy==1.19.4
fonttools==4.27.1
tqdm==4.54.1
